+++ crates/objc2/CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Added `ClassType::alloc_main_thread`.
 * Added `IsMainThreadOnly::mtm`.
 
+### Changed
+* Print backtrace when catching exceptions with the `"catch-all"` feature.
+
 
 ## 0.5.2 - 2024-05-21